requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
anthropic
langchain
langchain-anthropic
chromadb
sentence-transformers
requests
beautifulsoup4
playwright
python-jose
passlib
bcrypt
pydantic-settings
fastapi-mail